requirements.txt
asgiref==3.4.1
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==35.0.0
defusedxml==0.7.1
Django==3.0.5
django-allauth==0.50.0
django-cors-headers==3.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==3.3
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.20
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
